To Finish: Print, Restriction on Client Side, Registration, Account Recovery, flush

+++ public/admin/viewOrders.php
@@ -4,15 +4,16 @@
 include("./layouts/header.php");
 
 ?>
-<h2>Orders</h2>
-<hr>
 <div class="table-responsive p-3">
 
     <table id="dtMaterialDesignExample" class="table tabble-striped" cellspacing="0" width="100%">
         <thead class="blue white-text">
             <tr>
-                <th class="th-sm">Edit
-                </th>
+                <?php
+                restrict('<th class="th-sm">Edit
+        </th>');
+                ?>
+
                 <th class="th-sm">User ID
                 </th>
                 <th class="th-sm">Product ID
@@ -35,34 +36,40 @@ include("./layouts/header.php");
             <?php
             if ($result = getAllFetch("orders order by date desc")) {
                 while ($row = $result->fetch_assoc()) {
-                    $productExist = getById('products', $row['id']);
-
-
             ?>
                     <tr>
-                        <td>
-                            <a href="editOrder.php?orderID=<?php echo $row['id']
-                                                            ?>" class="btn btn-info btn-sm"><i class="fas fa-edit"></i></a>
-                        </td>
+
+                        <?php
+                        if ($row['state'] == 1) {
+                            echo '<td class="blue-text"><i class="fas fa-truck fa-lg"></i>';
+                        } else {
+
+                            restrict('<td><a href="editOrder.php?orderID=' . $row['date'] . '
+                                                            " class="btn btn-info btn-sm"><i class="fas fa-edit"></i></a>
+
+                        </td>');
+                        }
+
+                        ?>
+
                         <td><?php echo $row['userID'] ?></td>
                         <td><?php
-                            if ($productExist->num_rows == 0) {
-                                echo '<span class="red-text">' . $row['id'] . ' <i class="fas fa-database "></i> Product doesn\'t Exist Anymore</span>';
-                            } else {
-                                echo $row['id'];
+                            if ($productExist = getById('products', $row['id'], 1, "ID: " . $row['id'] . " Product Deleted")) {
+
+                                if ($productExist->num_rows == 0) {
+                                    echo '<span class="red-text">' . $row['id'] . ' <i class="fas fa-database "></i> Product doesn\'t Exist Anymore</span>';
+                                } else {
+                                    echo $row['id'];
+                                }
                             }
+
 
                             ?></td>
                         <td><?php echo $row['qty'] ?></td>
                         <td><?php echo $row['listPrice'] ?></td>
                         <td><?php
+                            displayAccountType($row['accountID']);
 
-
-                            if ($username = getById('accounts', $row['accountID'], $warning = 0)) {
-                                echo $username->fetch_assoc()['username'];
-                            } else {
-                                echo "<span class='red-text'>  <i class='fas fa-database'></i> Account ID: " . $row['accountID'] . " Account Deleted</span>";
-                            }
                             ?></td>
                         <?php echo orderState($row['state']);
                         ?>
